Refactor choice list classes.
