+++ src/Input/ChoiceList.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPForge\Html\Input;
+
+/*
+ * Generates a list of checkboxes.
+ *
+ * A checkbox or radio list allows many selections.
+ */
+final class ChoiceList extends Base\AbstractChoiceList
+{
+}
