+++ tests/Input/ChoiceList/ImmutabilityTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html\Tests\Input\CheckboxList;
+namespace PHPForge\Html\Tests\Input\ChoiceList;
 
 use PHPForge\Html\Input\Checkbox;
-use PHPForge\Html\Input\CheckboxList;
+use PHPForge\Html\Input\ChoiceList;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -15,7 +15,7 @@ final class ImmutabilityTest extends TestCase
 {
     public function testImmutability(): void
     {
-        $widget = CheckboxList::widget();
+        $widget = ChoiceList::widget();
 
         $this->assertNotSame($widget, $widget->items(Checkbox::widget()));
     }
